requirements.txt
requests==2.32.3
datetime==5.5
python-dotenv==1.0.1
datetime==5.5
telegram==0.0.1
python-telegram-bot==13.8